requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
pymongo==4.7.3
python-dotenv==1.0.1
bcrypt==4.1.3
flask-login==0.6.3
flask-wtf==1.2.1
wtforms==3.1.2
Faker==25.0.0
rich==13.7.1
tabulate==0.9.0
